修改gb28181库

+++ Channel/GB28181Channel/GB28181Option.cs
@@ -7,6 +7,6 @@ namespace GB28181Channel
         public string redis_conn { get; set; }
         public bool ipv6_enable { get; set; }
         public string sip_service_id { get; set; }
-        public int sip_listen_port { get; set; }
+        public ushort sip_listen_port { get; set; }
     }
 }
